+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[NEW]** Created new `MessagingSummaryItem`
 - [...]
 
 # v8.5.0 (09/07/2019)